requirements.txt
python-dotenv
discord
selenium
